requirements.txt
os
python-dotenv
Flask==2.2.2
line-bot-sdk
openai
